routine

+++ MyVSLib_C++/HybridC_v0.h
@@ -23,7 +23,7 @@ void fncElementWiseModSq(float* parrOut, complex<float>* pcarrInp, unsigned int 
 
 
 void fncCoherentDedispersion(complex<float>* pcarrCD_Out, complex<float>* pcarrffted_rowsignal
-	, const unsigned int LEnChunk, const float VAl_practicalD, const float VAlFmin, const float VAlFmax);
+	, const unsigned int LEnChunk, const long double VAl_practicalD, const float VAlFmin, const float VAlFmax);
 
 template <typename T>
 float fnsStdDev(T* parr_fdmt_inp, const float mean, unsigned int len);
